sentimind_project/sentimind_project/backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django-cors-headers>=4.9.0",
    "django-filter>=25.2",
    "djangorestframework>=3.16.1",
    "gunicorn>=23.0.0",
    "python-dotenv>=1.0.0",
    "transformers>=4.57.3",
    "torch",
    "sentencepiece>=0.2.0",
]

# Índice de PyTorch CPU-only (reduce de 2GB a 200MB)
[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true

[tool.uv.sources]
torch = { index = "pytorch-cpu" }

